requirements.txt
praw
PyYAML
python-dotenv
tqdm
langchain
sentence-transformers
chromadb
google-generativeai
